Validate SessionInfo in responses
Per the spec (https://github.com/teslamotors/vehicle-command/blob/main/pkg/protocol/protocol.md#recovering-from-synchronization-errors)

SessionInfo should be invalidated on an error if:
* The UUID doesn't match
* The HMAC tag does not match
* The clock time is out-of-order

This change removes the SessionInfo from the repository if any of these are true when a signing fault occurs, and performs a new handshake.

Additionally, if the session is valid, the session data is now updated from the responses. The counter is only updated if the epochs did not match.